requirements.txt

langgraph
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
tqdm
streamlit
